httpstatus상태를 변경하는데 post일때는 201 created put일때는 202 appcepted 를 구현하는 코드 작성

+++ src/main/kotlin/com/microservices/chapter3/CustomerController.kt
@@ -26,26 +26,45 @@ class CustomerController {
 
     @GetMapping(value = arrayOf("/customer/{id}"))
 //    fun getCustomer(@PathVariable id : Int) = customerService.getCustomer(id)
-    fun getCustomer(@PathVariable id : Int) =
-        ResponseEntity(customerService.getCustomer(id), HttpStatus.OK)
-
+    fun getCustomer(@PathVariable id : Int) : ResponseEntity<Customer?>{
+        val customer = customerService.getCustomer(id)
+        val status = if (customer == null) HttpStatus.NOT_FOUND else HttpStatus.OK
+        return ResponseEntity(customer, status)
+    }
     @GetMapping(value = arrayOf("/customers"))
     fun getCustomers(@RequestParam(required = false, defaultValue = "") nameFilter: String) =
         customerService.searchCustomer(nameFilter)
 
     @PostMapping(value = arrayOf("/customer"))
-    fun createCustomer(@RequestBody customer: Customer) {
+//    fun createCustomer(@RequestBody customer: Customer) {
+//        customerService.createCustomer(customer)
+//    }
+    fun createCustomer(@RequestBody customer: Customer) : ResponseEntity<Unit?>{
         customerService.createCustomer(customer)
+//        return ResponseEntity(Unit, HttpStatus.CREATED)
+        return ResponseEntity<Unit?>(null, HttpStatus.CREATED)
     }
 
-    @DeleteMapping(value = arrayOf("/custer/{id}"))
-    fun deleteCustomer(@PathVariable id: Int) {
-        customerService.deleteCustomer(id)
+    @DeleteMapping(value = arrayOf("/customer/{id}"))
+    fun deleteCustomer(@PathVariable id: Int) : ResponseEntity<Unit>{
+        var status = HttpStatus.NOT_FOUND
+        if (customerService.getCustomer(id) != null){
+            customerService.deleteCustomer(id)
+            status = HttpStatus.OK
+        }
+        return ResponseEntity(Unit, status)
     }
+
 
     @PutMapping(value = arrayOf("/customer/{id}"))
-    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer){
-        customerService.updateCustomer(id, customer)
+    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer) :
+            ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if ( customerService.getCustomer(id) != null){
+            customerService.updateCustomer(id, customer)
+            status = HttpStatus.ACCEPTED
+        }
+        return ResponseEntity(Unit, status)
     }
 }
